expose a way to save federation messages
