+++ ee/packages/federation-matrix/src/FederationMatrix.ts
@@ -1,4 +1,12 @@
-import { Authorization, type IFederationMatrixService, Room, ServiceClass, Settings } from '@rocket.chat/core-services';
+import {
+	Authorization,
+	type IFederationMatrixService,
+	Message,
+	MeteorService,
+	Room,
+	ServiceClass,
+	Settings,
+} from '@rocket.chat/core-services';
 import {
 	isDeletedMessage,
 	isMessageFromMatrixFederation,
@@ -9,13 +17,20 @@ import {
 } from '@rocket.chat/core-typings';
 import type { MessageQuoteAttachment, IMessage, IRoom, IUser, IRoomNativeFederated, ISubscription } from '@rocket.chat/core-typings';
 import { eventIdSchema, roomIdSchema, userIdSchema, federationSDK, FederationRequestError } from '@rocket.chat/federation-sdk';
-import type { EventID, FileMessageType, PresenceState } from '@rocket.chat/federation-sdk';
+import type { EventID, FileMessageType, PduForType, PresenceState } from '@rocket.chat/federation-sdk';
 import { Logger } from '@rocket.chat/logger';
 import { Users, Subscriptions, Messages, Rooms } from '@rocket.chat/models';
 
 import { createOrUpdateFederatedUser } from './helpers/createOrUpdateFederatedUser';
 import { extractDomainFromMatrixUserId } from './helpers/extractDomainFromMatrixUserId';
-import { toExternalMessageFormat, toExternalQuoteMessageFormat } from './helpers/message.parsers';
+import { getThreadMessageId } from './helpers/getThreadMessageId';
+import { handleMediaMessage } from './helpers/handleMediaMessage';
+import {
+	toExternalMessageFormat,
+	toExternalQuoteMessageFormat,
+	toInternalMessageFormat,
+	toInternalQuoteMessageFormat,
+} from './helpers/message.parsers';
 import { validateFederatedUsername } from './helpers/validateFederatedUsername';
 import { MatrixMediaService } from './services/MatrixMediaService';
 import { shortnameToUnicode } from './utils/emojiConverter';
@@ -1026,5 +1041,157 @@ export class FederationMatrix extends ServiceClass implements IFederationMatrixS
 				}
 			}),
 		);
+	}
+
+	async saveFederationMessage({ event, event_id: eventId }: { event: PduForType<'m.room.message'>; event_id: EventID }): Promise<void> {
+		const { msgtype, body } = event.content;
+		const messageBody = body.toString();
+
+		if (!messageBody && !msgtype) {
+			this.logger.debug('Received message event with empty body and no msgtype, skipping processing');
+			return;
+		}
+
+		// at this point we know for sure the user already exists
+		const user = await Users.findOneByUsername(event.sender);
+		if (!user) {
+			throw new Error(`User not found for sender: ${event.sender}`);
+		}
+
+		const room = await Rooms.findOne({ 'federation.mrid': event.room_id });
+		if (!room) {
+			throw new Error(`No mapped room found for room_id: ${event.room_id}`);
+		}
+
+		const serverName = federationSDK.getConfig('serverName');
+
+		const relation = event.content['m.relates_to'];
+
+		// SPEC: For example, an m.thread relationship type denotes that the event is part of a “thread” of messages and should be rendered as such.
+		const hasRelation = relation && 'rel_type' in relation;
+
+		const isThreadMessage = hasRelation && relation.rel_type === 'm.thread';
+
+		const threadRootEventId = isThreadMessage && relation.event_id;
+
+		// SPEC: Though rich replies form a relationship to another event, they do not use rel_type to create this relationship.
+		// Instead, a subkey named m.in_reply_to is used to describe the reply’s relationship,
+		const isRichReply = relation && !('rel_type' in relation) && 'm.in_reply_to' in relation;
+
+		const quoteMessageEventId = isRichReply && relation['m.in_reply_to']?.event_id;
+
+		const thread = threadRootEventId ? await getThreadMessageId(threadRootEventId) : undefined;
+
+		const isEditedMessage = hasRelation && relation.rel_type === 'm.replace';
+		if (isEditedMessage && relation.event_id && event.content['m.new_content']) {
+			this.logger.debug('Received edited message from Matrix, updating existing message');
+			const originalMessage = await Messages.findOneByFederationId(relation.event_id);
+			if (!originalMessage) {
+				this.logger.error({ event_id: relation.event_id, msg: 'Original message not found for edit' });
+				return;
+			}
+			if (originalMessage.federation?.eventId !== relation.event_id) {
+				return;
+			}
+			if (originalMessage.msg === event.content['m.new_content']?.body) {
+				this.logger.debug('No changes in message content, skipping update');
+				return;
+			}
+
+			if (quoteMessageEventId) {
+				const messageToReplyToUrl = await MeteorService.getMessageURLToReplyTo(room.t as string, room._id, originalMessage._id);
+				const formatted = await toInternalQuoteMessageFormat({
+					messageToReplyToUrl,
+					formattedMessage: event.content.formatted_body || '',
+					rawMessage: messageBody,
+					homeServerDomain: serverName,
+					senderExternalId: event.sender,
+				});
+				await Message.updateMessage(
+					{
+						...originalMessage,
+						msg: formatted,
+					},
+					user,
+					originalMessage,
+				);
+				return;
+			}
+
+			const formatted = toInternalMessageFormat({
+				rawMessage: event.content['m.new_content'].body,
+				formattedMessage: event.content.formatted_body || '',
+				homeServerDomain: serverName,
+				senderExternalId: event.sender,
+			});
+
+			await Message.updateMessage(
+				{
+					...originalMessage,
+					msg: formatted,
+				},
+				user,
+				originalMessage,
+			);
+			return;
+		}
+
+		if (quoteMessageEventId) {
+			const originalMessage = await Messages.findOneByFederationId(quoteMessageEventId);
+			if (!originalMessage) {
+				this.logger.error({ quoteMessageEventId, msg: 'Original message not found for quote' });
+				return;
+			}
+			const messageToReplyToUrl = await MeteorService.getMessageURLToReplyTo(room.t as string, room._id, originalMessage._id);
+			const formatted = await toInternalQuoteMessageFormat({
+				messageToReplyToUrl,
+				formattedMessage: event.content.formatted_body || '',
+				rawMessage: messageBody,
+				homeServerDomain: serverName,
+				senderExternalId: event.sender,
+			});
+			await Message.saveMessageFromFederation({
+				fromId: user._id,
+				rid: room._id,
+				msg: formatted,
+				federation_event_id: eventId,
+				thread,
+				ts: new Date(event.origin_server_ts),
+			});
+			return;
+		}
+
+		const isMediaMessage = Object.values(fileTypes).includes(msgtype as FileMessageType);
+		if (isMediaMessage && 'url' in event.content) {
+			const result = await handleMediaMessage(
+				event.content.url,
+				event.content.info,
+				msgtype,
+				messageBody,
+				user,
+				room,
+				event.room_id,
+				eventId,
+				thread,
+			);
+			await Message.saveMessageFromFederation({ ...result, ts: new Date(event.origin_server_ts) });
+			return;
+		}
+
+		const formatted = toInternalMessageFormat({
+			rawMessage: messageBody,
+			formattedMessage: event.content.formatted_body || '',
+			homeServerDomain: serverName,
+			senderExternalId: event.sender,
+		});
+
+		await Message.saveMessageFromFederation({
+			fromId: user._id,
+			rid: room._id,
+			msg: formatted,
+			federation_event_id: eventId,
+			thread,
+			ts: new Date(event.origin_server_ts),
+		});
 	}
 }
